Fix: 2D datetimeaxis rounding error
Datetimeaxis incorrectly rounded up sub-minute intervals. This happened
because qreal values were assigned to float. Usually qreal is double so
precission were lost in the type conversion. Change the code to use
qreal instead of float.

Pick-to: 6.12 6.11
Fixes: QTBUG-147722

+++ src/graphs2d/qsgrenderer/pointrenderer.cpp
@@ -171,8 +171,8 @@ void PointRenderer::calculateRenderCoordinates(AxisRenderer *axisRenderer,
     auto &axisX = axisRenderer->getAxisX(series);
     auto &axisY = axisRenderer->getAxisY(series);
 
-    float x = origX;
-    float y = origY;
+    qreal x = origX;
+    qreal y = origY;
 
     if (axisX.isLogarithmic) {
         float logBase = log(axisX.logBase);
